@@ -2,6 +2,9 @@ import { Component } from '@angular/core';
 import { Location } from '@angular/common';
 import { BaseMenuItem, Menus } from 'src/app/core';
 import { Store } from '@ngxs/store';
+import { MenuItemFormComponent } from '../../../shared/components/forms/menu-item-form/menu-item-form.component';
+import { ContentComponent } from '../../../shared/components/layouts/content/content.component';
+import { ViewComponent } from '../../../shared/components/layouts/view/view.component';
 
 const MenuItemPlaceholder: BaseMenuItem = {
 	name: '',
@@ -14,16 +17,18 @@ const MenuItemPlaceholder: BaseMenuItem = {
 };
 
 @Component({
-	selector: 'app-add-item',
-	templateUrl: './add-item.component.html',
-	styles: [
-		`
+    selector: 'app-add-item',
+    templateUrl: './add-item.component.html',
+    styles: [
+        `
 			:host {
 				width: 100%;
 				height: 100%;
 			}
 		`
-	]
+    ],
+    standalone: true,
+    imports: [ViewComponent, ContentComponent, MenuItemFormComponent]
 })
 export class AddItemComponent {
 	menuItem = MenuItemPlaceholder;
