@@ -1,20 +1,21 @@
-import { Component, Input } from '@angular/core';
-import { faUtensils } from '@fortawesome/free-solid-svg-icons';
+import { Component, Input } from "@angular/core";
+import { RouterLink } from "@angular/router";
 
 export interface IMenuItemCardConfig {
-	id: string;
-	name: string;
-	image: string;
-	tagline: string;
-	category: string;
+  id: string;
+  name: string;
+  image: string;
+  tagline: string;
+  category: string;
 }
 
 @Component({
-	selector: 'app-menu-item-card',
-	templateUrl: './menu-item-card.component.html',
-	styleUrls: ['./menu-item-card.component.scss']
+  selector: "app-menu-item-card",
+  templateUrl: "./menu-item-card.component.html",
+  styleUrls: ["./menu-item-card.component.scss"],
+  standalone: true,
+  imports: [RouterLink],
 })
 export class MenuItemCardComponent {
-	@Input() config!: IMenuItemCardConfig;
-	faUtensils = faUtensils;
+  @Input() config!: IMenuItemCardConfig;
 }
