@@ -1,9 +1,12 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { NgIf } from '@angular/common';
 
 @Component({
-	selector: 'app-content',
-	templateUrl: './content.component.html',
-	styleUrls: ['./content.component.scss']
+    selector: 'app-content',
+    templateUrl: './content.component.html',
+    styleUrls: ['./content.component.scss'],
+    standalone: true,
+    imports: [NgIf]
 })
 export class ContentComponent {
 	@Input() title: string = '';
