@@ -3,5 +3,5 @@ import { AuthConfig } from "./auth-config";
 export const environment = {
   production: true,
   auth: AuthConfig,
-  serverUrl: "http://localhost:7070",
+  serverUrl: "http://localhost:4040",
 };
